move api key/secret to state/base and cache base state instead
