+++ src/components/Auth/Auth.container.js
@@ -1,20 +1,21 @@
 import { connect } from 'react-redux'
-import actions from 'state/auth/actions'
+import authActions from 'state/auth/actions'
+import baseActions from 'state/base/actions'
 import Auth from './Auth'
 
 function mapStateToProps(state = {}) {
   return {
-    apiKey: state.auth.apiKey,
-    apiSecret: state.auth.apiSecret,
+    apiKey: state.base.apiKey,
+    apiSecret: state.base.apiSecret,
     isShown: state.auth.isShown,
     authStatus: state.auth.authStatus,
   }
 }
 
 const mapDispatchToProps = dispatch => ({
-  checkAuth: () => dispatch(actions.checkAuth()),
-  setKey: key => dispatch(actions.setApiKey(key)),
-  setSecret: secret => dispatch(actions.setApiSecret(secret)),
+  checkAuth: () => dispatch(authActions.checkAuth()),
+  setKey: key => dispatch(baseActions.setApiKey(key)),
+  setSecret: secret => dispatch(baseActions.setApiSecret(secret)),
 })
 
 const AuthContainer = connect(mapStateToProps, mapDispatchToProps)(Auth)
