+++ src/state/auth/actions.js
@@ -10,39 +10,6 @@ export function checkAuth() {
 }
 
 /**
- * Create an action to store API key.
- * @param {string} key
- */
-export function setApiKey(key) {
-  return {
-    type: types.SET_API_KEY,
-    payload: key,
-  }
-}
-
-/**
- * Create an action to store API secret.
- * @param {string} secret
- */
-export function setApiSecret(secret) {
-  return {
-    type: types.SET_API_SECRET,
-    payload: secret,
-  }
-}
-
-/**
- * Create an action to store Auth key (UUID).
- * @param {string} key
- */
-export function setAuthKey(key) {
-  return {
-    type: types.SET_AUTH_KEY,
-    payload: key,
-  }
-}
-
-/**
  * Create an action to show auth dialog.
  */
 export function showAuth() {
@@ -53,8 +20,5 @@ export function showAuth() {
 
 export default {
   checkAuth,
-  setApiKey,
-  setApiSecret,
-  setAuthKey,
   showAuth,
 }
