+++ src/state/auth/constants.js
@@ -1,8 +1,8 @@
 export default {
   CHECK_AUTH: 'BITFINEX/AUTH/CHECK_AUTH',
-  SET_API_KEY: 'BITFINEX/AUTH/SET_API_KEY',
-  SET_API_SECRET: 'BITFINEX/AUTH/SET_API_SECRET',
-  SET_AUTH_KEY: 'BITFINEX/AUTH/SET_AUTH_KEY',
+  // SET_API_KEY: 'BITFINEX/AUTH/SET_API_KEY', // moved to state/base
+  // SET_API_SECRET: 'BITFINEX/AUTH/SET_API_SECRET',
+  // SET_AUTH_KEY: 'BITFINEX/AUTH/SET_AUTH_KEY',
   SHOW_AUTH: 'BITFINEX/AUTH/SHOW_AUTH',
   UPDATE_AUTH_STATUS: 'BITFINEX/AUTH/UPDATE_AUTH_STATUS',
 }
