+++ src/state/auth/reducer.js
@@ -1,30 +1,15 @@
 import types from './constants'
 
 const initialState = {
-  apiKey: '',
-  apiSecret: '',
-  authKey: '',
+  // apiKey: '', // moved to state/base
+  // apiSecret: '',
+  // authKey: '',
   authStatus: null,
   isShown: true,
 }
 
 export function authReducer(state = initialState, action) {
   switch (action.type) {
-    case types.SET_API_KEY:
-      return {
-        ...state,
-        apiKey: action.payload,
-      }
-    case types.SET_API_SECRET:
-      return {
-        ...state,
-        apiSecret: action.payload,
-      }
-    case types.SET_AUTH_KEY:
-      return {
-        ...state,
-        authKey: action.payload,
-      }
     case types.UPDATE_AUTH_STATUS:
       return {
         ...state,
