+++ src/state/base/constants.js
@@ -1,0 +1,5 @@
+export default {
+  SET_API_KEY: 'BITFINEX/AUTH/SET_API_KEY',
+  SET_API_SECRET: 'BITFINEX/AUTH/SET_API_SECRET',
+  SET_AUTH_KEY: 'BITFINEX/AUTH/SET_AUTH_KEY',
+}
